updating to use ml4gw utils as much as possible
