+++ projects/sandbox/export/export/snapshotter.py
@@ -1,77 +1,14 @@
-from typing import TYPE_CHECKING, Optional, Tuple
+from typing import TYPE_CHECKING, Optional
 
-import torch
-
-from aframe.architectures.preprocessor import Whitener
+from aframe.architectures.preprocessor import (
+    BackgroundSnapshotter,
+    BatchWhitener,
+)
 from hermes.quiver import Platform
 from hermes.quiver.streaming import utils as streaming_utils
-from ml4gw.transforms import SpectralDensity
-from ml4gw.utils.slicing import unfold_windows
 
 if TYPE_CHECKING:
     from hermes.quiver.model import EnsembleModel, ExposedTensor
-
-
-class BackgroundSnapshotter(torch.nn.Module):
-    """Update a kernel with a new piece of streaming data"""
-
-    def __init__(
-        self,
-        psd_length,
-        kernel_length,
-        fduration,
-        sample_rate,
-        inference_sampling_rate,
-    ) -> None:
-        super().__init__()
-        state_length = kernel_length + fduration + psd_length
-        state_length -= 1 / inference_sampling_rate
-        self.state_size = int(state_length * sample_rate)
-
-    def forward(
-        self, update: torch.Tensor, snapshot: torch.Tensor
-    ) -> Tuple[torch.Tensor, ...]:
-        x = torch.cat([snapshot, update], axis=-1)
-        snapshot = x[:, :, -self.state_size :]
-        return x, snapshot
-
-
-class BatchWhitener(torch.nn.Module):
-    """Calculate the PSDs and whiten an entire batch of kernels at once"""
-
-    def __init__(
-        self,
-        kernel_length: float,
-        sample_rate: float,
-        inference_sampling_rate: float,
-        batch_size: int,
-        fduration: float,
-        fftlength: float,
-        highpass: Optional[float] = None,
-    ) -> None:
-        super().__init__()
-        self.stride_size = int(sample_rate / inference_sampling_rate)
-        self.kernel_size = int(kernel_length * sample_rate)
-
-        self.size = (batch_size - 1) * self.stride_size + self.kernel_size
-        self.size += int(fduration * sample_rate)
-
-        self.spectral_density = SpectralDensity(
-            sample_rate,
-            fftlength=fftlength,
-            overlap=None,
-            average="mean",
-            fast=highpass is not None,
-        )
-        self.whitener = Whitener(fduration, sample_rate, highpass)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        splits = [x.shape[-1] - self.size, self.size]
-        background, x = torch.split(x, splits, dim=-1)
-        psd = self.spectral_density(background.double())
-        x = self.whitener(x.double(), psd)
-        x = unfold_windows(x, self.kernel_size, self.stride_size)
-        return x[:, 0]
 
 
 def add_streaming_input_preprocessor(
